splited into two files for easier integration with other scripts
